+++ src/Provider/ACFProvider.php
@@ -60,87 +60,6 @@ class ACFProvider {
 
 
 	/**
-	 * Add settings button
-	 */
-	public function adminInit(){
-
-		if( !current_user_can('administrator') || WP_ENV != 'dev' )
-			return;
-
-		if( isset($_GET['clear_acf_meta']) )
-			$this->deleteUnusedMeta();
-
-		// Remove generated thumbnails option
-		add_settings_field('clean_unused_acf_meta', __('Advanced Custom Fields'), function(){
-
-			$unusedMeta = $this->getUnusedMeta();
-
-			if( $unusedMeta )
-				echo '<a class="button button-primary" href="'.get_admin_url().'?clear_acf_meta" title="Be carefull, fields must be synchronised">'.__('Remove').' '.$unusedMeta.' unused meta</a>';
-			else
-				echo __('Nothing to remove');
-
-		}, 'general');
-	}
-
-	/**
-	 * Clean acf meta
-	 */
-	public function deleteUnusedMeta(){
-
-		global $wpdb;
-
-		$deleteSql = "DELETE FROM `{$wpdb->prefix}postmeta` 
-	    WHERE `meta_key` IN 
-		( SELECT TRIM(LEADING '_' FROM `meta_key`) AS mk 
-			FROM (SELECT * FROM {$wpdb->prefix}postmeta) as pm
-			WHERE pm.`meta_value` regexp '^field_[0-9a-f]+' 
-				AND pm.`meta_value` NOT IN 
-					(SELECT `post_name` FROM `{$wpdb->prefix}posts` WHERE `post_type` = 'acf-field') 
-		) 
-		OR `meta_key` IN 
-		( SELECT `meta_key` AS mk 
-			FROM (SELECT * FROM {$wpdb->prefix}postmeta) as pm 
-			WHERE pm.`meta_value` regexp '^field_[0-9a-f]+' 
-				AND pm.`meta_value` NOT IN 
-					(SELECT `post_name` FROM `{$wpdb->prefix}posts` WHERE `post_type` = 'acf-field') 
-		)";
-
-		$wpdb->query($deleteSql);
-
-		wp_redirect( get_admin_url(null, 'options-general.php') );
-		exit;
-	}
-
-
-	/**
-	 * Count unused meta
-	 */
-	public function getUnusedMeta(){
-
-		global $wpdb;
-
-		$selectSql = "SELECT count(`meta_id`) FROM `{$wpdb->prefix}postmeta` 
-	    WHERE `meta_key` IN 
-		( SELECT TRIM(LEADING '_' FROM `meta_key`) AS mk 
-			FROM `{$wpdb->prefix}postmeta` 
-			WHERE `meta_value` regexp '^field_[0-9a-f]+' 
-				AND `meta_value` NOT IN 
-					(SELECT `post_name` FROM `{$wpdb->prefix}posts` WHERE `post_type` = 'acf-field') 
-		) 
-		OR `meta_key` IN 
-		( SELECT `meta_key` AS mk 
-			FROM `{$wpdb->prefix}postmeta` 
-			WHERE `meta_value` regexp '^field_[0-9a-f]+' 
-				AND `meta_value` NOT IN 
-					(SELECT `post_name` FROM `{$wpdb->prefix}posts` WHERE `post_type` = 'acf-field') 
-		)";
-
-		return $wpdb->get_var($selectSql);
-	}
-
-
-	/**
 	 * Customize basic toolbar
 	 * @param $toolbars
 	 * @return
@@ -187,7 +106,6 @@ class ACFProvider {
 			// Setup ACF Settings
 			add_action( 'acf/init', [$this, 'addSettings'] );
 			add_filter( 'acf/fields/wysiwyg/toolbars' , [$this, 'editToolbars']  );
-			add_action( 'admin_init', [$this, 'adminInit'] );
 			add_action( 'init', [$this, 'addOptionPages'] );
 		}
 	}