+++ src/Routing/permastructs.php
@@ -2,6 +2,7 @@
 
 namespace Metabolism\WordpressBundle;
 
+use Dflydev\DotAccessData\Data;
 use Symfony\Component\Routing\Route,
 	Symfony\Component\Routing\RouteCollection;
 
@@ -10,6 +11,12 @@ class Permastruct{
 	public $collection;
 	private $controller_name, $wp_rewrite, $locale;
 
+	/**
+	 * Permastruct constructor.
+	 * @param $collection
+	 * @param $locale
+	 * @param $controller_name
+	 */
 	public function __construct($collection, $locale, $controller_name)
 	{
 		global $wp_rewrite;
@@ -19,9 +26,24 @@ class Permastruct{
 		$this->locale = $locale;
 		$this->wp_rewrite = $wp_rewrite;
 
-		$this->addRoutes();
+		if( wp_maintenance_mode() )
+			$this->addMaintenanceRoute();
+		else
+			$this->addRoutes();
 	}
 
+	/**
+	 * Catch all url to display maintenance
+	 */
+	public function addMaintenanceRoute(){
+
+		$maintenanceController = $this->getControllerName('maintenance');
+		$this->addRoute('maintenance', '{req}', ['req'=>".*"], false, $maintenanceController);
+	}
+
+	/**
+	 * Define all routes from post types and taxonomies
+	 */
 	public function addRoutes(){
 
 		$this->addRoute('front', '');
@@ -124,6 +146,10 @@ class Permastruct{
 	}
 
 
+	/**
+	 * @param $name
+	 * @return string
+	 */
 	private function getControllerName( $name ){
 
 		$methodName = str_replace('_parent', '', $name);
@@ -132,6 +158,10 @@ class Permastruct{
 		return 'App\Controller\\'.$this->controller_name.'::'.$methodName.'Action';
 	}
 
+	/**
+	 * @param $struct
+	 * @return array
+	 */
 	private function getPaths( $struct ){
 
 		$path = str_replace('%/', '}/', str_replace('/%', '/{', $struct));
@@ -142,6 +172,13 @@ class Permastruct{
 		return ['singular'=>$path, 'archive'=>$path.'/'.$this->wp_rewrite->pagination_base.'/{page}'];
 	}
 
+	/**
+	 * @param $name
+	 * @param $struct
+	 * @param array $requirements
+	 * @param bool $paginate
+	 * @param bool $controllerName
+	 */
 	public function addRoute( $name, $struct, $requirements=[], $paginate=false, $controllerName=false )
 	{
 		$name = str_replace('_structure', '', $name);
@@ -150,11 +187,8 @@ class Permastruct{
 		$paths = $this->getPaths($struct);
 		$locale = $this->locale?'.'.$this->locale:'';
 
-		if( !empty($paths['singular']) || $name == 'front' ){
-
-			$route = new Route( $paths['singular'], ['_controller'=>$controller], $requirements);
-			$this->collection->add($name.$locale, $route);
-		}
+		$route = new Route( $paths['singular'], ['_controller'=>$controller], $requirements);
+		$this->collection->add($name.$locale, $route);
 
 		if( $paginate && !empty($paths['archive']) )
 		{
@@ -164,6 +198,7 @@ class Permastruct{
 	}
 }
 
+/** @var Data $_config */
 global $_config;
 $controller_name = $_config->get('extra_permastructs.controller', 'MainController');